requirements.txt
openai-whisper
torch
torchaudio
sounddevice
numpy
scipy
opencv-python
networkx
fastapi
uvicorn
